pyproject.toml
[tool.poetry]
name = "carbonne"
version = "0.1.0"
description = ""
authors = ["Houssem Gharsalli <houssem.gharsalli@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.31"
celery = "^5.4.0"
redis = "^5.0.8"
flower = "^2.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
